requirements.txt
appdirs==1.4.4
audioread==2.1.9
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
decorator==4.4.2
Flask==1.1.2
google==3.0.0
grpcio==1.35.0
grpcio-tools==1.35.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
librosa==0.8.0
llvmlite==0.35.0
MarkupSafe==1.1.1
numba==0.52.0
numpy==1.20.1
packaging==20.9
pooch==1.3.0
protobuf==3.15.1
pycparser==2.20
pyparsing==2.4.7
requests==2.25.1
resampy==0.2.2
scikit-learn==0.24.1
scipy==1.6.1
six==1.15.0
SoundFile==0.10.3.post1
soupsieve==2.2
SpeechRecognition==3.8.1
threadpoolctl==2.1.0
urllib3==1.26.3
Werkzeug==1.0.1
